Add fast hypervolume
