+++ docs/news.md
@@ -3,6 +3,11 @@
 This page documents notable changes introduced in each chronological release of the MOEA Framework.
 
 
+## Version 3.4 (TBD)
+
+  * New, fast implementation of the hypervolume calculation.
+
+
 ## Version 3.3 (10 Feb 2023)
 
   * Moves encode and decode methods to the Variable interface.
