+++ src/org/moeaframework/analysis/sensitivity/ExtractData.java
@@ -231,7 +231,7 @@ public class ExtractData extends CommandLineUtility {
 		Indicator indicator = null;
 		
 		if (option.equals("hypervolume")) {
-			indicator = new Hypervolume(problem, referenceSet);
+			indicator = Hypervolume.getInstance(problem, referenceSet);
 		} else if (option.equals("generational")) {
 			indicator = new GenerationalDistance(problem, referenceSet);
 		} else if (option.equals("inverted")) {
