+++ src/org/moeaframework/core/Settings.java
@@ -32,7 +32,6 @@ import java.util.Properties;
 import org.moeaframework.util.TypedProperties;
 import org.apache.commons.text.StringTokenizer;
 import org.moeaframework.core.NondominatedPopulation.DuplicateMode;
-import org.moeaframework.core.indicator.Hypervolume;
 
 /**
  * Global settings used by this framework.  The {@code PROPERTIES} object
