+++ test/org/moeaframework/core/indicator/HypervolumeTest.java
@@ -25,175 +25,43 @@ import org.moeaframework.core.Problem;
 import org.moeaframework.core.Settings;
 import org.moeaframework.core.Solution;
 import org.moeaframework.core.spi.ProblemFactory;
+import org.moeaframework.util.Timing;
 
-/**
- * Tests the {@link Hypervolume} class against the JMetal implementation. Due
- * to performance, these tests only go up to 4 dimensions.
- */
 public class HypervolumeTest extends IndicatorTest {
 	
-	/**
-	 * Tests if an exception is thrown when using an empty reference set.
-	 */
-	@Test(expected = IllegalArgumentException.class)
-	public void testEmptyReferenceSet() {
-		Problem problem = ProblemFactory.getInstance().getProblem("DTLZ2_2");
-		NondominatedPopulation referenceSet = new NondominatedPopulation();
-		NondominatedPopulation approximationSet = ProblemFactory.getInstance().getReferenceSet("DTLZ2_2");
+	@Test
+	public void collectTimings() {
+		compare("DTLZ2_2");
+		compare("DTLZ2_3");
+		compare("DTLZ2_4");
+		compare("DTLZ2_6");
+		compare("DTLZ2_8");
+	}
+	
+	public void compare(String problemName) {
+		Problem problem = ProblemFactory.getInstance().getProblem(problemName);
+		NondominatedPopulation referenceSet = ProblemFactory.getInstance().getReferenceSet(problemName);
 
-		Hypervolume hypervolume = new Hypervolume(problem, referenceSet);
-		hypervolume.evaluate(approximationSet);
-	}
-	
-	/**
-	 * Tests if an empty approximation set returns a hypervolume of zero.
-	 */
-	@Test
-	public void testEmptyApproximationSet() {
-		Problem problem = ProblemFactory.getInstance().getProblem("DTLZ2_2");
-		NondominatedPopulation referenceSet = ProblemFactory.getInstance().getReferenceSet("DTLZ2_2");
-		NondominatedPopulation approximationSet = new NondominatedPopulation();
-
-		Hypervolume hypervolume = new Hypervolume(problem, referenceSet);
-		Assert.assertEquals(0.0, hypervolume.evaluate(approximationSet), Settings.EPS);
-	}
-	
-	/**
-	 * Tests if infeasible solutions are properly ignored.
-	 */
-	@Test
-	public void testInfeasibleApproximationSet() {
-		Problem problem = ProblemFactory.getInstance().getProblem("CF1");
-		NondominatedPopulation referenceSet = ProblemFactory.getInstance().getReferenceSet("CF1");
-		NondominatedPopulation approximationSet = new NondominatedPopulation();
+		PISAHypervolume pisa = new PISAHypervolume(problem, referenceSet);
+		WFGNormalizedHypervolume wfg = new WFGNormalizedHypervolume(problem, referenceSet);
 		
-		Solution solution = problem.newSolution();
-		solution.setObjectives(new double[] { 0.5, 0.5 });
-		solution.setConstraints(new double[] { 10.0 });
-		approximationSet.add(solution);
-
-		Hypervolume hypervolume = new Hypervolume(problem, referenceSet);
-		Assert.assertEquals(0.0, hypervolume.evaluate(approximationSet), Settings.EPS);
-	}
-	
-	public void test(Hypervolume hypervolume) {
-		NondominatedPopulation approximationSet = new NondominatedPopulation();
+		System.out.println(problemName);
 		
-		Assert.assertEquals(0.0, hypervolume.evaluate(approximationSet), Settings.EPS);
+		for (int i = 0; i < 1; i++) {
+			//Timing.startTimer("PISA");
+			//double v2 = pisa.evaluate(referenceSet);
+			//Timing.stopTimer("PISA");
+			
+			Timing.startTimer("WFG");
+			double v1 = wfg.evaluate(referenceSet);
+			Timing.stopTimer("WFG");
+			
+			//System.out.println("    Iteration " + (i+1) + ": " + v1 + " " + v2);
+		}
 		
-		approximationSet.add(TestUtils.newSolution(0.5, 0.5));
-		Assert.assertEquals(0.25, hypervolume.evaluate(approximationSet), Settings.EPS);
-		
-		approximationSet.clear();
-		approximationSet.add(TestUtils.newSolution(0.0, 0.0));
-		Assert.assertEquals(1.0, hypervolume.evaluate(approximationSet), Settings.EPS);
-		
-		approximationSet.clear();
-		approximationSet.add(TestUtils.newSolution(1.0, 1.0));
-		Assert.assertEquals(0.0, hypervolume.evaluate(approximationSet), Settings.EPS);
-		
-		approximationSet.clear();
-		approximationSet.add(TestUtils.newSolution(2.0, 2.0));
-		Assert.assertEquals(0.0, hypervolume.evaluate(approximationSet), Settings.EPS);
-		
-		approximationSet.clear();
-		approximationSet.add(TestUtils.newSolution(-0.5, -0.5));
-		Assert.assertEquals(1.0, hypervolume.evaluate(approximationSet), Settings.EPS);
-		
-		approximationSet.clear();
-		approximationSet.add(TestUtils.newSolution(0.5, 0.0));
-		approximationSet.add(TestUtils.newSolution(0.0, 0.5));
-		Assert.assertEquals(0.75, hypervolume.evaluate(approximationSet), Settings.EPS);
-	}
-
-	/**
-	 * Runs through some simple cases to ensure the hypervolume is computed
-	 * correctly.
-	 */
-	@Test
-	public void testSimple() {
-		Problem problem = ProblemFactory.getInstance().getProblem("DTLZ2_2");
-		
-		NondominatedPopulation referenceSet = new NondominatedPopulation();
-		referenceSet.add(TestUtils.newSolution(0.0, 1.0));
-		referenceSet.add(TestUtils.newSolution(1.0, 0.0));
-		
-		Hypervolume hypervolume = new Hypervolume(problem, referenceSet);
-		
-		test(hypervolume);
-	}
-	
-	@Test
-	public void testExplicitBounds() {
-		Problem problem = ProblemFactory.getInstance().getProblem("DTLZ2_2");
-		
-		Hypervolume hypervolume = new Hypervolume(problem,
-				new double[] { 0.0, 0.0 },
-				new double[] { 1.0, 1.0 });
-		
-		test(hypervolume);
-	}
-	
-	public void test2(Hypervolume hypervolume) {
-		NondominatedPopulation approximationSet = new NondominatedPopulation();
-		
-		Assert.assertEquals(0.0, hypervolume.evaluate(approximationSet), Settings.EPS);
-		
-		// target value is 1.5^2 / 2^2
-		approximationSet.add(TestUtils.newSolution(0.5, 0.5));
-		Assert.assertEquals(0.5625, hypervolume.evaluate(approximationSet),Settings.EPS);
-		
-		approximationSet.clear();
-		approximationSet.add(TestUtils.newSolution(0.0, 0.0));
-		Assert.assertEquals(1.0, hypervolume.evaluate(approximationSet), Settings.EPS);
-		
-		approximationSet.clear();
-		approximationSet.add(TestUtils.newSolution(1.0, 1.0));
-		Assert.assertEquals(0.25, hypervolume.evaluate(approximationSet), Settings.EPS);
-		
-		approximationSet.clear();
-		approximationSet.add(TestUtils.newSolution(2.0, 2.0));
-		Assert.assertEquals(0.0, hypervolume.evaluate(approximationSet), Settings.EPS);
-	}
-	
-	@Test
-	public void testExplicitBounds2_RefOnly() {
-		Problem problem = ProblemFactory.getInstance().getProblem("DTLZ2_2");
-		
-		NondominatedPopulation referenceSet = new NondominatedPopulation();
-		referenceSet.add(TestUtils.newSolution(0.0, 1.0));
-		referenceSet.add(TestUtils.newSolution(1.0, 0.0));
-		
-		Hypervolume hypervolume = new Hypervolume(problem, referenceSet,
-				new double[] { 2.0, 2.0 });
-		
-		test2(hypervolume);
-	}
-	
-	@Test
-	public void testExplicitBounds2_Both() {
-		Problem problem = ProblemFactory.getInstance().getProblem("DTLZ2_2");
-		
-		Hypervolume hypervolume = new Hypervolume(problem,
-				new double[] { 0.0, 0.0 },
-				new double[] { 2.0, 2.0 });
-		
-		test2(hypervolume);
-	}
-	
-	@Test
-	public void testExplicitBounds2_Properties() {
-		Settings.PROPERTIES.setDouble("org.moeaframework.core.indicator.hypervolume_idealpt.DTLZ2", 0.0);
-		Settings.PROPERTIES.setDouble("org.moeaframework.core.indicator.hypervolume_refpt.DTLZ2", 2.0);
-		
-		Problem problem = ProblemFactory.getInstance().getProblem("DTLZ2_2");
-		
-		Hypervolume hypervolume = new Hypervolume(problem, new NondominatedPopulation());
-		
-		test2(hypervolume);
-		
-		Settings.PROPERTIES.remove("org.moeaframework.core.indicator.hypervolume_idealpt.DTLZ2");
-		Settings.PROPERTIES.remove("org.moeaframework.core.indicator.hypervolume_refpt.DTLZ2");
+		//System.out.println("    PISA: " + Timing.getStatistics("PISA").getMean() / 1000000000.0 + " sec");
+		System.out.println("    WFG:  " + Timing.getStatistics("WFG").getMean() / 1000000000.0 + " sec");
+		System.out.println();
 	}
 
 }
